Initial Commit
new file:   Makefile
new file:   Project.sh
new file:   res/login.glade
new file:   src/main.c

Created new Project for a password manager written in C.
It use GTK GUIs designed in Glade.

Makefile created to help compile and run the program.
Makefile at this point, is not functioning.

Project.sh created as a way to store repeated bash commands that the
Mankefile Can't handle.

login.glade is the login gui for the password manager.
It allows the user to enter in their username and password.

main.c is the main source file for this C project.
At this point it contains code to connect to the login.glade file and
display it.
It also allows the glade file to be terminated